TSR-001: Bernard Walter Premier Freeman.
At this point Paul Hayes has signed an NDA with JUKERSTONE.
The isolation of this app from the network repo is (1) it was not a monorepo, but an amalgamation of projects, (2) for github actions, ci/cd a simple repo structure seems simplest to handle.

Aim is to achieve operational excellence in this repository.